requirements.txt
tradingview-screener>=3.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
google-genai>=1.0.0
google-adk>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.40
ta-lib>=0.6.4
